+++ src/TinyIRReceiver.cpp.h
@@ -45,6 +45,11 @@
 
 TinyIRReceiverStruct TinyIRReceiverControl;
 
+#if defined(ESP8266)
+ICACHE_RAM_ATTR
+#elif defined(ESP32)
+IRAM_ATTR
+#endif
 void handleReceivedIRData(uint16_t aAddress, uint8_t aCommand, bool isRepetition);
 
 bool MatchDuration(uint16_t aDuration, uint16_t aLowerMatchValue, uint16_t aUpperMatchValue) {
@@ -52,13 +57,11 @@ bool MatchDuration(uint16_t aDuration, uint16_t aLowerMatchValue, uint16_t aUppe
 }
 
 #if defined(ESP8266)
-void ICACHE_RAM_ATTR IRPinChangeInterruptHandler(void)
+ICACHE_RAM_ATTR
 #elif defined(ESP32)
-void IRAM_ATTR IRPinChangeInterruptHandler(void)
-#else
-void IRPinChangeInterruptHandler(void)
+IRAM_ATTR
 #endif
-        {
+void IRPinChangeInterruptHandler(void) {
     // save IR input level - negative logic, true means inactive / IR pause
     uint_fast8_t tIRLevel = digitalReadFast(IR_INPUT_PIN);